feat(sdk,skill): auto-align overlay by anchor coordinate calculation
SDK: change offsetX/offsetY from Int to Float across WebViewState,
ImageState, WebViewRenderer, ImageRenderer, InspectorPanel, DomQueryService
and InspectorApiHandler so sub-dp precision is preserved end-to-end.

Skill: rewrite inspect phase-1 from manual user alignment to automatic
anchor-based offset calculation — push_html → dom_all + get_all_nodes →
compute offsetX/offsetY from text-matched anchor → adjust_overlay.
Human only needs to navigate to the target page.

+++ packages/android/sdk/src/main/kotlin/com/clienttools/sdk/inspector/DomQueryService.kt
@@ -55,8 +55,8 @@ class DomQueryService(
 
     suspend fun queryAll(
         webView: WebView,
-        webViewOffsetXDp: Int,
-        webViewOffsetYDp: Int
+        webViewOffsetXDp: Float,
+        webViewOffsetYDp: Float
     ): List<DomNodeInfo> {
         val density = webView.context.resources.displayMetrics.density
         val offsetXPx = (webViewOffsetXDp * density).toInt()
@@ -92,8 +92,8 @@ class DomQueryService(
     suspend fun queryById(
         webView: WebView,
         id: String,
-        webViewOffsetXDp: Int,
-        webViewOffsetYDp: Int
+        webViewOffsetXDp: Float,
+        webViewOffsetYDp: Float
     ): DomNodeInfo? {
         val density = webView.context.resources.displayMetrics.density
         val offsetXPx = (webViewOffsetXDp * density).toInt()
