+++ packages/android/sdk/src/main/kotlin/com/clienttools/sdk/inspector/WebViewRenderer.kt
@@ -56,9 +56,9 @@ class WebViewRenderer(rootView: View, private val viewModel: InspectorViewModel)
         job = null
     }
 
-    private fun dpToPx(dp: Int): Float = TypedValue.applyDimension(
+    private fun dpToPx(dp: Float): Float = TypedValue.applyDimension(
         TypedValue.COMPLEX_UNIT_DIP,
-        dp.toFloat(),
+        dp,
         webView.context.resources.displayMetrics
     )
 }
